Separate into class
